ModelGroupWriter now works with the ResourceLoader and ResourceHandler

+++ webtools/webapp/entity/ModelGroupWriter.jsp
@@ -1,4 +1,4 @@
-<%@ page contentType="text/plain" %><%@ page import="java.util.*, java.io.*, java.net.*, org.ofbiz.core.util.*, org.ofbiz.core.entity.*, org.ofbiz.core.entity.model.*" %><jsp:useBean id="delegator" type="org.ofbiz.core.entity.GenericDelegator" scope="application" /><jsp:useBean id="security" type="org.ofbiz.core.security.Security" scope="application" /><%
+<%@ page contentType="text/plain" %><%@ page import="java.util.*, java.io.*, java.net.*, org.ofbiz.core.util.*, org.ofbiz.core.entity.*, org.ofbiz.core.entity.config.*, org.ofbiz.core.entity.model.*" %><jsp:useBean id="delegator" type="org.ofbiz.core.entity.GenericDelegator" scope="application" /><jsp:useBean id="security" type="org.ofbiz.core.security.Security" scope="application" /><%
 
   if("true".equals(request.getParameter("savetofile"))) {
     if(security.hasPermission("ENTITY_MAINT", session)) {
@@ -7,36 +7,38 @@
       String serverRootUrl=(String)request.getAttribute(SiteDefs.SERVER_ROOT_URL);
       ModelGroupReader modelGroupReader = delegator.getModelGroupReader();
 
-      String filename = modelGroupReader.entityGroupFileName;
+      ResourceHandler resourceHandler = modelGroupReader.entityGroupResourceHandler;
+      if (resourceHandler.isFileResource()) {
+          String filename = resourceHandler.getFullLocation();
 
-      java.net.URL url = new java.net.URL(serverRootUrl + controlPath + "/view/ModelGroupWriter");
-      HashMap params = new HashMap();
-      HttpClient httpClient = new HttpClient(url, params);
-      InputStream in = httpClient.getStream();
+          java.net.URL url = new java.net.URL(serverRootUrl + controlPath + "/view/ModelGroupWriter");
+          HashMap params = new HashMap();
+          HttpClient httpClient = new HttpClient(url, params);
+          InputStream in = httpClient.getStream();
 
-      File newFile = new File(filename);
-      FileWriter newFileWriter = new FileWriter(newFile);
+          File newFile = new File(filename);
+          FileWriter newFileWriter = new FileWriter(newFile);
 
-      BufferedReader post = new BufferedReader(new InputStreamReader(in));
-      String line = null;
-      while((line = post.readLine()) != null) {
-        newFileWriter.write(line);
-        newFileWriter.write("\n");
-      }
-      newFileWriter.close();
-      %>
-      If you aren't seeing any exceptions, XML was written successfully to:
-      <%=filename%>
-      from the URL:
-      <%=url.toString()%>
+          BufferedReader post = new BufferedReader(new InputStreamReader(in));
+          String line = null;
+          while((line = post.readLine()) != null) {
+            newFileWriter.write(line);
+            newFileWriter.write("\n");
+          }
+          newFileWriter.close();
+          %>
+          If you aren't seeing any exceptions, XML was written successfully to:
+          <%=filename%>
+          from the URL:
+          <%=url.toString()%>
       <%
-    } 
-    else {
+      } else {
+        %>ERROR: This entity group information did not come from a file, so it is not being saved. It came from <%=resourceHandler.toString()%><%
+      }
+    } else {
       %>ERROR: You do not have permission to use this page (ENTITY_MAINT needed)<%
     }
-  }
-  else
-  {
+  } else {
 %><?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE entitygroup PUBLIC "-//OFBiz//DTD Entity Group//EN" "http://www.ofbiz.org/dtds/entitygroup.dtd">
 <!--
